Jan 29, 2026, 3:13 PM

+++ DAA/Assignment_1/binary_search.py
@@ -9,7 +9,7 @@ left = 0
 right = n-1
 f = 0
 while(left <= right):
-    mid = left + (right - left) / 2
+    mid = int(left + (right - left) / 2)
     if arr[mid] == search_ele:
         f = 1
         break
